+++ src/main/java/com/epam/reportportal/cucumber/AbstractReporter.java
@@ -28,16 +28,18 @@ import io.cucumber.plugin.ConcurrentEventListener;
 import io.cucumber.plugin.event.*;
 import io.reactivex.Maybe;
 import org.apache.commons.lang3.tuple.Pair;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import rp.com.google.common.base.Supplier;
 import rp.com.google.common.base.Suppliers;
 import rp.com.google.common.io.ByteSource;
 
 import java.net.URI;
-import java.util.*;
+import java.util.Calendar;
+import java.util.Date;
+import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
+import static com.epam.reportportal.cucumber.Utils.getCodeRef;
+import static com.epam.reportportal.cucumber.Utils.getDescription;
 import static rp.com.google.common.base.Strings.isNullOrEmpty;
 import static rp.com.google.common.base.Throwables.getStackTraceAsString;
 
@@ -54,11 +56,9 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 
 	private static final String AGENT_PROPERTIES_FILE = "agent.properties";
 
-    protected Supplier<Launch> launch;
+	protected Supplier<Launch> launch;
 	static final String COLON_INFIX = ": ";
 	private static final String SKIPPED_ISSUE_KEY = "skippedIssue";
-	private static final Logger LOGGER = LoggerFactory.getLogger(AbstractReporter.class);
-
 	private final Map<URI, RunningContext.FeatureContext> currentFeatureContextMap = new ConcurrentHashMap<>();
 
 	private final Map<Pair<String, URI>, RunningContext.ScenarioContext> currentScenarioContextMap = new ConcurrentHashMap<>();
@@ -68,11 +68,7 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 	// End of feature occurs once launch is finished.
 	private final Map<URI, Date> featureEndTime = new ConcurrentHashMap<>();
 
-	private Map<Long, RunningContext.ScenarioContext> threadCurrentScenarioContextMap = new ConcurrentHashMap<>();
-
-	protected void setThreadCurrentScenarioContextMap(Map<Long, RunningContext.ScenarioContext> threadCurrentScenarioContextMap) {
-		this.threadCurrentScenarioContextMap = threadCurrentScenarioContextMap;
-	}
+	private final Map<Long, RunningContext.ScenarioContext> threadCurrentScenarioContextMap = new ConcurrentHashMap<>();
 
 	/**
 	 * Registers an event handler for a specific event.
@@ -136,13 +132,15 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 	/**
 	 * Start Cucumber scenario
 	 */
-	protected void beforeScenario(RunningContext.FeatureContext currentFeatureContext, RunningContext.ScenarioContext currentScenarioContext,
-			String scenarioName) {
-		Maybe<String> id = Utils.startNonLeafNode(
-				launch.get(),
+	protected void beforeScenario(RunningContext.FeatureContext currentFeatureContext,
+			RunningContext.ScenarioContext currentScenarioContext, String scenarioName) {
+		String description = getDescription(currentFeatureContext.getUri());
+		String codeRef = getCodeRef(currentFeatureContext.getUri(), currentScenarioContext.getLine());
+		Maybe<String> id = Utils.startNonLeafNode(launch.get(),
 				currentFeatureContext.getFeatureId(),
 				scenarioName,
-				currentFeatureContext.getUri() + ":" + currentScenarioContext.getLine(),
+				description,
+				codeRef,
 				currentScenarioContext.getAttributes(),
 				getScenarioTestItemType()
 		);
@@ -275,8 +273,7 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 	}
 
 	protected void embedding(String mimeType, byte[] data) {
-		ReportPortal.emitLog(
-				new ReportPortalMessage(ByteSource.wrap(data), mimeType, mimeType),
+		ReportPortal.emitLog(new ReportPortalMessage(ByteSource.wrap(data), mimeType, mimeType),
 				"UNKNOWN",
 				Calendar.getInstance().getTime()
 		);
@@ -292,46 +289,63 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 
 	protected abstract Maybe<String> getRootItemId();
 
+	private RunningContext.FeatureContext createFeatureContext(TestCase testCase) {
+		RunningContext.FeatureContext currentFeatureContext;
+		currentFeatureContext = new RunningContext.FeatureContext().processTestSourceReadEvent(testCase);
+		String featureKeyword = currentFeatureContext.getFeature().getKeyword();
+		String featureName = currentFeatureContext.getFeature().getName();
+		StartTestItemRQ rq = new StartTestItemRQ();
+		Maybe<String> root = getRootItemId();
+		rq.setDescription(getDescription(currentFeatureContext.getUri()));
+		rq.setCodeRef(getCodeRef(currentFeatureContext.getUri(), 0));
+		rq.setName(Utils.buildNodeName(featureKeyword, AbstractReporter.COLON_INFIX, featureName, null));
+		rq.setAttributes(currentFeatureContext.getAttributes());
+		rq.setStartTime(Calendar.getInstance().getTime());
+		rq.setType(getFeatureTestItemType());
+		currentFeatureContext.setFeatureId(root == null ? launch.get().startTestItem(rq) : launch.get().startTestItem(root, rq));
+		return currentFeatureContext;
+	}
+
 	/**
 	 * Private part that responsible for handling events
 	 */
 
-    protected EventHandler<TestRunStarted> getTestRunStartedHandler() {
+	protected EventHandler<TestRunStarted> getTestRunStartedHandler() {
 		return event -> beforeLaunch();
 	}
 
-    protected EventHandler<TestSourceRead> getTestSourceReadHandler() {
+	protected EventHandler<TestSourceRead> getTestSourceReadHandler() {
 		return event -> RunningContext.FeatureContext.addTestSourceReadEvent(event.getUri(), event);
 	}
 
-    protected EventHandler<TestCaseStarted> getTestCaseStartedHandler() {
+	protected EventHandler<TestCaseStarted> getTestCaseStartedHandler() {
 		return this::handleStartOfTestCase;
 	}
 
-    protected EventHandler<TestStepStarted> getTestStepStartedHandler() {
+	protected EventHandler<TestStepStarted> getTestStepStartedHandler() {
 		return this::handleTestStepStarted;
 	}
 
-    protected EventHandler<TestStepFinished> getTestStepFinishedHandler() {
+	protected EventHandler<TestStepFinished> getTestStepFinishedHandler() {
 		return this::handleTestStepFinished;
 	}
 
-    protected EventHandler<TestCaseFinished> getTestCaseFinishedHandler() {
+	protected EventHandler<TestCaseFinished> getTestCaseFinishedHandler() {
 		return this::afterScenario;
 	}
 
-    protected EventHandler<TestRunFinished> getTestRunFinishedHandler() {
+	protected EventHandler<TestRunFinished> getTestRunFinishedHandler() {
 		return event -> {
 			handleEndOfFeature();
 			afterLaunch();
 		};
 	}
 
-    protected EventHandler<EmbedEvent> getEmbedEventHandler() {
+	protected EventHandler<EmbedEvent> getEmbedEventHandler() {
 		return event -> embedding(event.getMediaType(), event.getData());
 	}
 
-    protected EventHandler<WriteEvent> getWriteEventHandler() {
+	protected EventHandler<WriteEvent> getWriteEventHandler() {
 		return event -> write(event.getText());
 	}
 
@@ -347,17 +361,16 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 		TestCase testCase = event.getTestCase();
 		RunningContext.FeatureContext featureContext = new RunningContext.FeatureContext().processTestSourceReadEvent(testCase);
 		URI featureUri = featureContext.getUri();
-		RunningContext.FeatureContext currentFeatureContext = currentFeatureContextMap.get(featureUri);
-
-		currentFeatureContext = currentFeatureContext == null ? createFeatureContext(testCase, featureUri) : currentFeatureContext;
+		RunningContext.FeatureContext currentFeatureContext = currentFeatureContextMap.computeIfAbsent(featureUri,
+				u -> createFeatureContext(testCase)
+		);
 
 		if (!currentFeatureContext.getUri().equals(testCase.getUri())) {
 			throw new IllegalStateException("Scenario URI does not match Feature URI.");
 		}
 
 		RunningContext.ScenarioContext scenarioContext = currentFeatureContext.getScenarioContext(testCase);
-		String scenarioName = Utils.buildNodeName(
-				scenarioContext.getKeyword(),
+		String scenarioName = Utils.buildNodeName(scenarioContext.getKeyword(),
 				AbstractReporter.COLON_INFIX,
 				scenarioContext.getName(),
 				scenarioContext.getOutlineIteration()
@@ -373,24 +386,6 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 		}
 
 		beforeScenario(currentFeatureContext, currentScenarioContext, scenarioName);
-	}
-
-	private RunningContext.FeatureContext createFeatureContext(TestCase testCase, URI featureURI) {
-		RunningContext.FeatureContext currentFeatureContext;
-		currentFeatureContext = new RunningContext.FeatureContext().processTestSourceReadEvent(testCase);
-		currentFeatureContextMap.put(featureURI, currentFeatureContext);
-		String featureKeyword = currentFeatureContext.getFeature().getKeyword();
-		String featureName = currentFeatureContext.getFeature().getName();
-
-		StartTestItemRQ rq = new StartTestItemRQ();
-		Maybe<String> root = getRootItemId();
-		rq.setDescription(currentFeatureContext.getUri().toString());
-		rq.setName(Utils.buildNodeName(featureKeyword, AbstractReporter.COLON_INFIX, featureName, null));
-		rq.setAttributes(currentFeatureContext.getAttributes());
-		rq.setStartTime(Calendar.getInstance().getTime());
-		rq.setType(getFeatureTestItemType());
-		currentFeatureContext.setFeatureId(root == null ? launch.get().startTestItem(rq) : launch.get().startTestItem(root, rq));
-		return currentFeatureContext;
 	}
 
 	protected void handleTestStepStarted(TestStepStarted event) {