+++ src/test/java/com/epam/reportportal/cucumber/LaunchSystemAttributesTest.java
@@ -6,17 +6,12 @@ import com.epam.reportportal.service.ReportPortalClient;
 import com.epam.ta.reportportal.ws.model.attribute.ItemAttributesRQ;
 import com.epam.ta.reportportal.ws.model.launch.StartLaunchRQ;
 import io.reactivex.Maybe;
-import io.reactivex.MaybeEmitter;
-import io.reactivex.MaybeOnSubscribe;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
-import org.mockito.invocation.InvocationOnMock;
-import org.mockito.stubbing.Answer;
 import rp.com.google.common.collect.Lists;
 
 import java.util.HashMap;
@@ -24,7 +19,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.regex.Pattern;
 
-import static org.mockito.Matchers.any;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.*;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
 /**
@@ -32,7 +29,7 @@ import static org.mockito.Mockito.*;
  */
 public class LaunchSystemAttributesTest {
 
-	private static final Map<String, Pattern> properties = new HashMap<String, Pattern>();
+	private static final Map<String, Pattern> properties = new HashMap<>();
 
 	private static final String SKIPPED_ISSUE_KEY = "skippedIssue";
 
@@ -44,18 +41,17 @@ public class LaunchSystemAttributesTest {
 	@Mock
 	private ListenerParameters listenerParameters;
 
-	@BeforeClass
+	@BeforeAll
 	public static void initKeys() {
 		properties.put("os", Pattern.compile("^.+\\|.+\\|.+$"));
 		properties.put("jvm", Pattern.compile("^.+\\|.+\\|.+$"));
 		properties.put("agent", Pattern.compile("^test-agent\\|test-1\\.0$"));
 	}
 
-	@Before
+	@BeforeEach
 	public void initLaunch() {
 		MockitoAnnotations.initMocks(this);
 		when(listenerParameters.getEnable()).thenReturn(true);
-		when(listenerParameters.getBaseUrl()).thenReturn("http://example.com");
 		when(listenerParameters.getIoPoolSize()).thenReturn(10);
 		when(listenerParameters.getBatchLogsSize()).thenReturn(5);
 		stepReporter = new StepReporter() {
@@ -68,18 +64,10 @@ public class LaunchSystemAttributesTest {
 
 	@Test
 	public void shouldRetrieveSystemAttributes() {
-		when(reportPortalClient.startLaunch(any(StartLaunchRQ.class))).then(new Answer<Object>() {
-			@Override
-			public Object answer(InvocationOnMock t) throws Throwable {
-				return Maybe.create(new MaybeOnSubscribe<Object>() {
-					@Override
-					public void subscribe(MaybeEmitter<Object> emitter) throws Exception {
-						emitter.onSuccess("launchId");
-						emitter.onComplete();
-					}
-				}).cache();
-			}
-		});
+		when(reportPortalClient.startLaunch(any(StartLaunchRQ.class))).then(t -> Maybe.create(emitter -> {
+			emitter.onSuccess("launchId");
+			emitter.onComplete();
+		}).cache());
 
 		stepReporter.beforeLaunch();
 
@@ -90,7 +78,7 @@ public class LaunchSystemAttributesTest {
 
 		StartLaunchRQ startLaunchRequest = launchRQArgumentCaptor.getValue();
 
-		Assert.assertNotNull(startLaunchRequest.getAttributes());
+		assertThat(startLaunchRequest.getAttributes(), notNullValue());
 
 		List<ItemAttributesRQ> attributes = Lists.newArrayList(startLaunchRequest.getAttributes());
 
@@ -100,21 +88,17 @@ public class LaunchSystemAttributesTest {
 			}
 		}
 
-		Assert.assertEquals(3, attributes.size());
-
-		for (ItemAttributesRQ attribute : attributes) {
-			Assert.assertTrue(attribute.isSystem());
-
-			Pattern pattern = LaunchSystemAttributesTest.this.getPattern(attribute);
-			Assert.assertNotNull(pattern);
-			Assert.assertTrue(pattern.matcher(attribute.getValue()).matches());
-		}
-
+		assertThat(attributes, hasSize(3));
+		attributes.forEach(attr -> {
+			assertThat(attr.isSystem(), equalTo(Boolean.TRUE));
+			Pattern pattern = LaunchSystemAttributesTest.this.getPattern(attr);
+			assertThat(pattern, notNullValue());
+			assertThat(attr.getValue(), allOf(notNullValue(), matchesPattern(pattern)));
+		});
 	}
 
 	private Pattern getPattern(ItemAttributesRQ attribute) {
 		return properties.get(attribute.getKey());
 
 	}
-
 }