+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Test Case ID support
+### Fixed
+- codeRef reporting was added for every item in an item tree
 
 ## [0.0.1-ALPHA]
-
-## [0.0.1]
 ## Added
+- Initial release to Public Maven Repositories